+++ react/__mocks__/vtex.styleguide.tsx
@@ -1,7 +1,5 @@
-import Dropdown from '@vtex/styleguide/lib/Dropdown'
-import Input from '@vtex/styleguide/lib/Input'
-import Link from '@vtex/styleguide/lib/Link'
-import Spinner from '@vtex/styleguide/lib/Spinner'
-import IconDelete from '@vtex/styleguide/lib/icon/Delete'
-
-module.exports = { Input, Dropdown, Link, Spinner, IconDelete }
+export { default as Dropdown } from '@vtex/styleguide/lib/Dropdown'
+export { default as Input } from '@vtex/styleguide/lib/Input'
+export { default as Link } from '@vtex/styleguide/lib/Link'
+export { default as Spinner } from '@vtex/styleguide/lib/Spinner'
+export { default as IconDelete } from '@vtex/styleguide/lib/icon/Delete'